pyproject.toml
[project]
name = "checkpointflow"
version = "1.0.0"
description = "CLI toolchain for authoring and running deterministic, resumable agent workflows."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.26.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
cpf = "checkpointflow.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "PTH", "RET", "RUF", "SIM", "UP"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.13"
strict = true
packages = ["checkpointflow"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
addopts = "-ra --maxfail=1"
testpaths = ["tests"]
